[UPDATE] views/homepage.notifications.show
improved ux

+++ resources/views/homepage/notifications/show.blade.php
@@ -123,22 +123,28 @@
                                 </td>
                                 <td>
                                     @if ($notice['deleted_at'] == null)
-                                        <a class="text-primary mx-1" data-bs-toggle="tooltip" title="Edit"
-                                            href="{{ route('homepage.notification.editPage', $notice['id']) }}">
-                                            <span class="material-icons">edit</span></a>
-                                        <a class="text-danger mx-1" data-bs-toggle="tooltip" title="Delete"
-                                            href="{{ route('homepage.notification.softDelete', $notice['id']) }}"
-                                            alert-title="Move to Trash?" alert-text="-"
-                                            confirm spoof spoof-method="DELETE"><span class="material-icons">delete</span></a>
+                                        @if (Auth::user()->can('update', \App\Models\HomepageNotification::class))
+                                            <a class="text-primary mx-1" data-bs-toggle="tooltip" title="Edit"
+                                                href="{{ route('homepage.notification.editPage', $notice['id']) }}">
+                                                <span class="material-icons">edit</span></a>
+                                            <a class="text-danger mx-1" data-bs-toggle="tooltip" title="Delete"
+                                                href="{{ route('homepage.notification.softDelete', $notice['id']) }}"
+                                                alert-title="Move to Trash?" alert-text="-"
+                                                confirm spoof spoof-method="DELETE"><span class="material-icons">delete</span></a>
+                                        @endif
                                     @else
-                                        <a class="text-success mx-1" data-bs-toggle="tooltip" title="Restore"
-                                            href="{{ route('homepage.notification.restore', $notice['id']) }}"
-                                            spoof spoof-method="POST">
-                                            <span class="material-icons">restore</span></a>
-                                        <a class="text-danger mx-1" data-bs-toggle="tooltip" title="Delete Permanently"
-                                            href="{{ route('homepage.notification.delete', $notice['id']) }}"
-                                            alert-title="Delete Permanently?" confirm spoof spoof-method="DELETE">
-                                            <span class="material-icons">delete_forever</span></a>
+                                        @if (Auth::user()->can('update', \App\Models\HomepageNotification::class))
+                                            <a class="text-success mx-1" data-bs-toggle="tooltip" title="Restore"
+                                                href="{{ route('homepage.notification.restore', $notice['id']) }}"
+                                                spoof spoof-method="POST">
+                                                <span class="material-icons">restore</span></a>
+                                        @endif
+                                        @if (Auth::user()->can('delete', \App\Models\HomepageNotification::class))
+                                            <a class="text-danger mx-1" data-bs-toggle="tooltip" title="Delete Permanently"
+                                                href="{{ route('homepage.notification.delete', $notice['id']) }}"
+                                                alert-title="Delete Permanently?" confirm spoof spoof-method="DELETE">
+                                                <span class="material-icons">delete_forever</span></a>
+                                        @endif
                                     @endif
                                 </td>
                             </tr>
